+++ changelog.md
@@ -17,6 +17,7 @@
     - All armor and weapon objects can now have custom skill requirements set for all or per different classes
     - Food and generic objects can be turned into fishing bait
     - Add room and object read support
+    - Merge all the No base * in to single property field to make grids cleaner
   - Add [#setsetting](docs/commands.md) options for chat, see [faq](docs/faq.md) for list of all supported settings
   - Add [#chat](docs/commands.md) command to send test to chat window and append newline
   - Add [#chatprompt](docs/commands.md) command to send test to chat window with no newline